+++ verik-compiler/regression/07-imported.sv
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Francis Wang
+ * Copyright (c) 2022 Francis Wang
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,20 @@
  * limitations under the License.
  */
 
-package io.verik.compiler.ast.property
+// Test.sv /////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-enum class PackageType {
-    REGULAR_NON_ROOT,
-    REGULAR_ROOT,
-    IMPORTED_NON_ROOT,
-    IMPORTED_ROOT;
+`ifndef VERIK
+`define VERIK
+`timescale 1ns / 1ns
+`endif
 
-    fun isRoot(): Boolean {
-        return this in listOf(REGULAR_ROOT, IMPORTED_ROOT)
-    }
+module M;
 
-    fun isImported(): Boolean {
-        return this in listOf(IMPORTED_NON_ROOT, IMPORTED_ROOT)
-    }
-}
+    initial begin : f
+        logic x1;
+        C     c;
+        x1 = x0;
+        c = C::new();
+    end : f
+
+endmodule : M