+++ verik-compiler/src/main/kotlin/io/verik/compiler/serialize/source/SourceSerializerVisitor.kt
@@ -42,7 +42,6 @@ import io.verik.compiler.ast.element.expression.common.ECallExpression
 import io.verik.compiler.ast.element.expression.common.EConstantExpression
 import io.verik.compiler.ast.element.expression.common.EExpression
 import io.verik.compiler.ast.element.expression.common.EIfExpression
-import io.verik.compiler.ast.element.expression.common.ENullExpression
 import io.verik.compiler.ast.element.expression.common.EParenthesizedExpression
 import io.verik.compiler.ast.element.expression.common.EPropertyStatement
 import io.verik.compiler.ast.element.expression.common.EReferenceExpression
@@ -237,10 +236,6 @@ class SourceSerializerVisitor(
 
     override fun visitStructLiteralExpression(structLiteralExpression: EStructLiteralExpression) {
         ExpressionSerializer.serializeStructLiteralExpression(structLiteralExpression, serializeContext)
-    }
-
-    override fun visitNullExpression(nullExpression: ENullExpression) {
-        ExpressionSerializer.serializeNullExpression(serializeContext)
     }
 
     override fun visitThisExpression(thisExpression: EThisExpression) {