+++ verik-compiler/src/test/kotlin/io/verik/compiler/transform/upper/IfAndWhenExpressionUnlifterStageTest.kt
@@ -49,7 +49,7 @@ internal class IfAndWhenExpressionUnlifterStageTest : BaseTest() {
                     y = __0;
                 endfunction : f
             """.trimIndent()
-        ) { it.regularPackageTextFiles[0] }
+        ) { it.nonRootPackageTextFiles[0] }
     }
 
     @Test
@@ -81,7 +81,7 @@ internal class IfAndWhenExpressionUnlifterStageTest : BaseTest() {
                     y = __0;
                 endfunction : f
             """.trimIndent()
-        ) { it.regularPackageTextFiles[0] }
+        ) { it.nonRootPackageTextFiles[0] }
     }
 
     @Test
@@ -114,6 +114,6 @@ internal class IfAndWhenExpressionUnlifterStageTest : BaseTest() {
                     y = __0;
                 endfunction : f
             """.trimIndent()
-        ) { it.regularPackageTextFiles[0] }
+        ) { it.nonRootPackageTextFiles[0] }
     }
 }